refactor: rename the app package to config and move the global exception handler #307
