+++ backend/src/main/java/com/occupi/config/CacheConfig.java
@@ -1,4 +1,4 @@
-package com.occupi.app;
+package com.occupi.config;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
 import org.springframework.beans.factory.annotation.Value;
